Refactor inputs to do their own insert()-ing
